% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Actor}
\alias{Actor}
\title{Base R6 Actor Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Actor
}
\description{
This \link[R6:R6Class]{R6 class} is a constructor for all Actor subclasses.
Objects are instantiated with the following methods:
\itemize{
\item A list of \emph{character parameters} (\code{params})
\item A method for attacking another object (\code{\link[=attack]{attack()}})
\item A method for healing another object (\code{\link[=heal]{heal()}})
}
}
\keyword{datasets}
