% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{actor_classes}
\alias{actor_classes}
\alias{starting_classes}
\alias{Actor}
\title{Base R6 Actor Class}
\format{An object of class \code{list} of length 3.}
\usage{
starting_classes

Actor
}
\description{
This \link[R6:R6Class]{R6 class} is a constructor for all Actor subclasses. Also
documented here are various related R objects, like a list of possible
starting classes (Which means you need to document better!)
}
\keyword{datasets}
