% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{actor_jobs}
\alias{actor_jobs}
\title{Player Classes}
\format{An object of class \code{list} of length 3.}
\usage{
actor_jobs
}
\description{
A list of possible Actor starting jobs.
}
\keyword{datasets}
