% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat.R
\name{attack_}
\alias{attack_}
\title{Attack}
\usage{
attack_(attacker, target)
}
\arguments{
\item{attacker}{\link{Actor}-class or subclass instance that is attacking}

\item{target}{\link{Actor}-class or subclass that is the target of the attack}
}
\description{
This function causes HP damage to another \link{Actor}-class instance. The damage
dealt is relative to the attacker's Strength parameter, \code{str} (specifically:
Damage = (5 * str)).
}
