% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat.R
\name{heal}
\alias{heal}
\title{Heal}
\usage{
heal(caster, target)
}
\arguments{
\item{caster}{\link{Actor}-class or subclass instance that is casting}

\item{target}{\link{Actor}-class or subclass that is the target of the cast}
}
\description{
This function restores HP to another \link{Actor}-class instance (or to self). The
HP restored is relative to the attacker's Magicka parameter, \code{mgk}.
}
