% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{main_menu}
\alias{main_menu}
\alias{help}
\alias{start_quest}
\title{Menu Functions}
\usage{
help()

start_quest()
}
\description{
These functions are available to the player before beginning the game itself.
}
