% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Player}
\alias{Player}
\title{Player Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Player
}
\description{
R6 class generator for Player Classes. Serves to instantiate a Player object.
Inherits from \link{Actor}. Add params to initialize() that re-set ceiling for HP
and MP restoration.
}
\keyword{datasets}
