% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{player_classes}
\alias{player_classes}
\title{Player Classes}
\format{An object of class \code{list} of length 3.}
\usage{
player_classes
}
\description{
A list of possible player starting classes.
}
\keyword{datasets}
