% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrcov.R
\name{qrcov}
\alias{qrcov}
\title{Approximate Covariance Matrix Estimation for Ratios of Linear Combinations of Quantile Estimators}
\usage{
qrcov(x, u, coef1, coef2, quantile.type = 8, ...)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{u}{a numeric vector of probability values in the interval (0,1) specifying the quantiles to be estimated. Note that u must include numeric values between, and not including, 0 and 1 and missing values are not allowed.}

\item{coef1}{a numeric matrix whose \eqn{i}th row contains the coefficients for the numerator linear combination for the \eqn{i}th ratio.}

\item{coef2}{a numeric matrix whose \eqn{i}th row contains the coefficients for the denominator linear combination for the \eqn{i}th ratio.}

\item{quantile.type}{argument for the quantile function.  Default is set to 8 so that output is consistent with default quantile function use and other functions such as IQR (see help file for \code{quantile()}
for more details)}

\item{...}{additional arguments to be passed to function density when method = “density” is used.}
}
\value{
a list consisting of \code{ratios} (the ratio estimates) and \code{cov} (the estimated covariance matrix).
}
\description{
compute a covariance matrix consisting of variances (on the diagonal) for ratios of linear combinations of quantile estimates
and covariances (off-diagonal) between different ratios
}
\details{
This function computes a sample covariance matrix for ratios of linear combinations of quantile estimates from a single sample.
This is done via a sample covariance matrix for the quantiles used in the linear combinations which is obtained using function \code{qcov}.
Let \eqn{\mathbf{q}} denote a vector of quantile estimators, \eqn{\mathbf{S}} denote the covariance matrix for \eqn{\mathbf{q}},
and \eqn{\mathbf{c}_{1i}} and \eqn{\mathbf{c}_{2i}} \eqn{i=1,\ldots,r} be the numerator and denominator vectors of coefficients
that define the linear combinations used in each of the \eqn{r} ratios (e.g. the \eqn{i}th ratio is
\eqn{R_i = (\mathbf{c}_{1i}^\top \mathbf{q})/(\mathbf{c}_{2i}^\top \mathbf{q})}).  Then, using the Delta method, an approximate
covariance matrix for \eqn{[R_1,\ldots,R_r]^\top} is
\deqn{\mathbf{J}\mathbf{S}\mathbf{J}^\top} where \eqn{\mathbf{J}} is the Jacobian matrix whose \eqn{i}th row is the gradient
vector of \eqn{R_i}.
}
\examples{
# Create some data
set.seed(1234)
x <- rnorm(100)

# Compute the variance-covariance matrix for the ratio of third
# quartile and the median and the ratio of first quartile and the median.
coef1 <- matrix(c(0, 0, 1, 1, 0, 0), nrow = 2, byrow = TRUE)
coef2 <- matrix(c(0, 1, 0, 0, 1, 0), nrow = 2, byrow = TRUE)
qrcov(x, c(0.25, 0.5, 0.75), coef1 = coef1, coef2 = coef2)
}
\references{
Prendergast, L. A., Dedduwakumara, D.S. & Staudte, R.G. (2024) rquest: An R package for hypothesis tests and confidence intervals
for quantiles and summary measures based on quantiles, preprint, pages 1-13
}
