% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat.R
\name{restore_mp_}
\alias{restore_mp_}
\title{Restore MP to Target}
\usage{
restore_mp_(caster, target)
}
\description{
Restore MP to target through concentration. Costs no MP to use, but drains 5%
of max health from caster. The MP restored is relative to the attacker's
Magicka parameter, \code{mgk} (specifically: MP restored = (2 * mgk)).
}
