% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Villain}
\alias{Villain}
\title{Villain Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Villain
}
\description{
R6 class generator for Villain Classes. Serves to instantiate a Villain object.
Inherits from \link{Actor}. Add params to initialize() that set ceiling for HP and
MP restoration.
}
\keyword{datasets}
