% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquest-package.R
\docType{package}
\name{rquest-package}
\alias{rquest}
\alias{rquest-package}
\title{rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles}
\description{
The rquest package provides a simple way to conduct hypothesis tests and derive confidence intervals for quantiles, linear combinations of quantiles, ratios of dependent linear combinations (e.g., Bowley’s measure of skewness) and differences and ratios of all of the above for comparisons between independent samples. Additionally, quantile-based measures of inequality are also considered.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shenal-dkumara/rquest}
  \item Report bugs at \url{https://github.com/shenal-dkumara/rquest/issues}
}

}
\author{
\strong{Maintainer}: Shenal Dedduwakumara \email{shenal.dedduwakumara@adelaide.edu.au}

Authors:
\itemize{
  \item Luke Prendergast \email{luke.prendergast@latrobe.edu.au}
  \item Robert Staudte \email{r.staudte@latrobe.edu.au}
}

}
\keyword{internal}
